+++ src/main/java/tudelft/ti2806/pl3/visualization/FilteredGraphModel.java
@@ -8,9 +8,10 @@ import tudelft.ti2806.pl3.data.wrapper.WrappedGraphData;
 import tudelft.ti2806.pl3.data.wrapper.Wrapper;
 import tudelft.ti2806.pl3.data.wrapper.operation.interest.CalculateSizeInterest;
 import tudelft.ti2806.pl3.data.wrapper.operation.yposition.PositionNodeYOnGenomeSpace;
-import tudelft.ti2806.pl3.util.wrap.WrapUtil;
+import tudelft.ti2806.pl3.data.wrapper.util.WrapUtil;
 //import tudelft.ti2806.pl3.visualization.wrapper.operation.interest.CalculateAddMaxOfWrapped;
 //import tudelft.ti2806.pl3.visualization.wrapper.operation.interest.CalculateWrapPressureInterest;
+
 
 
 import java.util.ArrayList;
